+++ Packages/PaintKit/Sources/PaintDemo/DemoScene.swift
@@ -5,12 +5,14 @@ enum DemoScene: String, CaseIterable, Sendable {
     case clipboard
     case quickSketch = "quick-sketch"
     case transparency
+    case chameleon
 
     var defaultFilename: String {
         switch self {
         case .clipboard: "clipboard-markup.png"
         case .quickSketch: "quick-sketch.png"
         case .transparency: "transparency.png"
+        case .chameleon: "chameleon.png"
         }
     }
 
@@ -23,6 +25,8 @@ enum DemoScene: String, CaseIterable, Sendable {
             QuickSketchScene.render()
         case .transparency:
             TransparencyScene.render()
+        case .chameleon:
+            ChameleonScene.render()
         }
     }